enhancement: allow user to visit RadioGroupContext
